Move FilenameService to Core
